reset resty client timeout
